Able to now append registered players to game page
I have figured out how to get it working. It really just came down to finding the reservations data in the API then going through it to find the info I wanted. Though this works, there is certainly a more efficient of doing it but that will be for the next commit ;)